% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affinity.R
\name{affinity_supported}
\alias{affinity_supported}
\title{Check whether CPU affinity is supported}
\usage{
affinity_supported()
}
\value{
A logical scalar indicating platform support.
}
\description{
Currently supported on Linux only.
}
\examples{
affinity_supported()
}
