% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer.R
\name{as.array.shard_buffer}
\alias{as.array.shard_buffer}
\title{Coerce a Shared Memory Buffer to Array}
\usage{
\method{as.array}{shard_buffer}(x, ...)
}
\arguments{
\item{x}{A \code{shard_buffer} object.}

\item{...}{Ignored.}
}
\value{
An array with the buffer contents and the buffer's dimensions, or a
plain vector for 1-D buffers.
}
\description{
Coerce a Shared Memory Buffer to Array
}
\examples{
\donttest{
buf <- buffer("double", dim = c(2, 3, 4))
as.array(buf)
buffer_close(buf)
}
}
