% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer.R
\name{as.double.shard_buffer}
\alias{as.double.shard_buffer}
\title{Coerce a Shared Memory Buffer to Double}
\usage{
\method{as.double}{shard_buffer}(x, ...)
}
\arguments{
\item{x}{A \code{shard_buffer} object.}

\item{...}{Ignored.}
}
\value{
A double vector with the buffer contents.
}
\description{
Coerce a Shared Memory Buffer to Double
}
\examples{
\donttest{
buf <- buffer("double", dim = 5)
as.double(buf)
buffer_close(buf)
}
}
