% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer.R
\name{as.integer.shard_buffer}
\alias{as.integer.shard_buffer}
\title{Coerce a Shared Memory Buffer to Integer}
\usage{
\method{as.integer}{shard_buffer}(x, ...)
}
\arguments{
\item{x}{A \code{shard_buffer} object.}

\item{...}{Ignored.}
}
\value{
An integer vector with the buffer contents.
}
\description{
Coerce a Shared Memory Buffer to Integer
}
\examples{
\donttest{
buf <- buffer("integer", dim = 5)
as.integer(buf)
buffer_close(buf)
}
}
