% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer.R
\name{as.logical.shard_buffer}
\alias{as.logical.shard_buffer}
\title{Coerce a Shared Memory Buffer to Logical}
\usage{
\method{as.logical}{shard_buffer}(x, ...)
}
\arguments{
\item{x}{A \code{shard_buffer} object.}

\item{...}{Ignored.}
}
\value{
A logical vector with the buffer contents.
}
\description{
Coerce a Shared Memory Buffer to Logical
}
\examples{
\donttest{
buf <- buffer("logical", dim = 5)
as.logical(buf)
buffer_close(buf)
}
}
