% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer.R
\name{as.matrix.shard_buffer}
\alias{as.matrix.shard_buffer}
\title{Coerce a Shared Memory Buffer to Matrix}
\usage{
\method{as.matrix}{shard_buffer}(x, ...)
}
\arguments{
\item{x}{A \code{shard_buffer} object (must be 2-dimensional).}

\item{...}{Ignored.}
}
\value{
A matrix with the buffer contents and the buffer's dimensions.
}
\description{
Coerce a Shared Memory Buffer to Matrix
}
\examples{
\donttest{
buf <- buffer("double", dim = c(3, 4))
as.matrix(buf)
buffer_close(buf)
}
}
