% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer.R
\name{as.raw.shard_buffer}
\alias{as.raw.shard_buffer}
\title{Coerce a Shared Memory Buffer to Raw}
\usage{
\method{as.raw}{shard_buffer}(x, ...)
}
\arguments{
\item{x}{A \code{shard_buffer} object.}

\item{...}{Ignored.}
}
\value{
A raw vector with the buffer contents.
}
\description{
Coerce a Shared Memory Buffer to Raw
}
\examples{
\donttest{
buf <- buffer("raw", dim = 5)
as.raw(buf)
buffer_close(buf)
}
}
