% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer.R
\name{as.vector.shard_buffer}
\alias{as.vector.shard_buffer}
\title{Coerce a Shared Memory Buffer to a Vector}
\usage{
\method{as.vector}{shard_buffer}(x, mode = "any")
}
\arguments{
\item{x}{A \code{shard_buffer} object.}

\item{mode}{Storage mode passed to \code{\link{as.vector}}.}
}
\value{
A vector of the buffer's type (or coerced to \code{mode}).
}
\description{
Coerce a Shared Memory Buffer to a Vector
}
\examples{
\donttest{
buf <- buffer("double", dim = 5)
buf[1:5] <- 1:5
as.vector(buf)
buffer_close(buf)
}
}
