% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{as_tibble.shard_dataset}
\alias{as_tibble.shard_dataset}
\title{Materialize a dataset handle into a data.frame/tibble}
\usage{
\method{as_tibble}{shard_dataset}(x, max_bytes = 256 * 1024^2, ...)
}
\arguments{
\item{x}{A \code{shard_dataset} handle.}

\item{max_bytes}{Accepted for API consistency.}

\item{...}{Reserved for future extensions.}
}
\value{
A data.frame (or tibble if the \code{tibble} package is installed).
}
\description{
Materialize a dataset handle into a data.frame/tibble
}
