% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{as_tibble.shard_table_handle}
\alias{as_tibble.shard_table_handle}
\title{Materialize a table handle into a data.frame/tibble}
\usage{
\method{as_tibble}{shard_table_handle}(x, max_bytes = 256 * 1024^2, ...)
}
\arguments{
\item{x}{A \code{shard_table_handle}.}

\item{max_bytes}{Warn if estimated payload exceeds this threshold.}

\item{...}{Reserved for future extensions.}
}
\value{
A data.frame (or tibble if the \code{tibble} package is installed).
}
\description{
Materialize a table handle into a data.frame/tibble
}
