% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assert_positive_int}
\alias{assert_positive_int}
\title{Validate Positive Integer}
\usage{
assert_positive_int(x, name = "value")
}
\arguments{
\item{x}{Value to validate.}

\item{name}{Name for error messages.}
}
\value{
Integer value.
}
\description{
Validate Positive Integer
}
\keyword{internal}
