% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cow.R
\name{attr<-.shard_shared_vector}
\alias{attr<-.shard_shared_vector}
\title{Set an Attribute on a Shared Vector}
\usage{
\method{attr}{shard_shared_vector}(x, which) <- value
}
\arguments{
\item{x}{A \code{shard_shared_vector}.}

\item{which}{Attribute name.}

\item{value}{Attribute value.}
}
\value{
The modified object \code{x}.
}
\description{
Raises an error if the copy-on-write policy is \code{"deny"}.
}
