% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cow.R
\name{attributes<-.shard_shared_vector}
\alias{attributes<-.shard_shared_vector}
\title{Set Attributes on a Shared Vector}
\usage{
\method{attributes}{shard_shared_vector}(x) <- value
}
\arguments{
\item{x}{A \code{shard_shared_vector}.}

\item{value}{Named list of attributes.}
}
\value{
The modified object \code{x}.
}
\description{
Raises an error if the copy-on-write policy is \code{"deny"}.
}
