% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{available_backings}
\alias{available_backings}
\title{Get available shared memory backing types}
\usage{
available_backings()
}
\value{
A character vector of available backing types on the current platform.
}
\description{
Get available shared memory backing types
}
\examples{
available_backings()
}
