% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affinity.R
\name{buffer_advise}
\alias{buffer_advise}
\title{Advise access pattern for a buffer}
\usage{
buffer_advise(
  x,
  advice = c("normal", "sequential", "random", "willneed", "dontneed")
)
}
\arguments{
\item{x}{A shard_buffer.}

\item{advice}{See \code{\link[=segment_advise]{segment_advise()}}.}
}
\value{
A logical scalar; \code{TRUE} if the OS accepted the hint.
}
\description{
Advise access pattern for a buffer
}
\examples{
\donttest{
buf <- buffer("double", dim = 10L)
buffer_advise(buf, "sequential")
}
}
