% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer.R
\name{buffer_close}
\alias{buffer_close}
\title{Close a Buffer}
\usage{
buffer_close(x, unlink = NULL)
}
\arguments{
\item{x}{A shard_buffer object.}

\item{unlink}{Whether to unlink the underlying segment.}
}
\value{
\code{NULL}, invisibly.
}
\description{
Closes the buffer and releases the underlying shared memory.
}
\examples{
\donttest{
buf <- buffer("double", dim = 10)
buffer_close(buf)
}
}
