% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer.R
\name{buffer_path}
\alias{buffer_path}
\title{Get Buffer Path}
\usage{
buffer_path(x)
}
\arguments{
\item{x}{A shard_buffer object.}
}
\value{
A character string with the path or name of the segment, or
\code{NULL} if the segment is anonymous.
}
\description{
Returns the path or name of the buffer's underlying segment.
Use this to pass buffer location to workers.
}
\examples{
\donttest{
buf <- buffer("double", dim = 10)
buffer_path(buf)
buffer_close(buf)
}
}
