% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{capture_time}
\alias{capture_time}
\title{Capture Time}
\usage{
capture_time(expr)
}
\arguments{
\item{expr}{Expression to time.}
}
\value{
List with result and elapsed time.
}
\description{
Simple timing wrapper.
}
\keyword{internal}
