% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/share.R
\name{close.shard_shared}
\alias{close.shard_shared}
\alias{close.shard_shared_vector}
\alias{close.shard_deep_shared}
\title{Close a Shared Object}
\usage{
\method{close}{shard_shared}(con, ...)

\method{close}{shard_shared_vector}(con, ...)

\method{close}{shard_deep_shared}(con, ...)
}
\arguments{
\item{con}{A \code{shard_shared} object.}

\item{...}{Ignored.}
}
\value{
NULL (invisibly).
}
\description{
Releases the shared memory segment. After closing, the shared object can
no longer be accessed.
}
