% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{collect.shard_dataset}
\alias{collect.shard_dataset}
\title{Collect a dataset handle into memory}
\usage{
\method{collect}{shard_dataset}(x, ...)
}
\arguments{
\item{x}{A \code{shard_dataset} handle.}

\item{...}{Passed to \code{as_tibble()}.}
}
\value{
A data.frame (or tibble if the \code{tibble} package is installed).
}
\description{
Collect a dataset handle into memory
}
