% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{collect.shard_row_groups}
\alias{collect.shard_row_groups}
\title{Collect a row-groups handle into memory}
\usage{
\method{collect}{shard_row_groups}(x, ...)
}
\arguments{
\item{x}{A \code{shard_row_groups} handle.}

\item{...}{Passed to \code{as_tibble()}.}
}
\value{
A data.frame (or tibble if the \code{tibble} package is installed).
}
\description{
Collect a row-groups handle into memory
}
