% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{collect.shard_table_handle}
\alias{collect.shard_table_handle}
\title{Collect a table handle into memory}
\usage{
\method{collect}{shard_table_handle}(x, ...)
}
\arguments{
\item{x}{A \code{shard_table_handle}.}

\item{...}{Passed to \code{as_tibble()}.}
}
\value{
A data.frame (or tibble if the \code{tibble} package is installed).
}
\description{
Collect a table handle into memory
}
