% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{coltypes}
\alias{coltypes}
\alias{int32}
\alias{float64}
\alias{bool}
\alias{raw_col}
\alias{string_col}
\title{Column Types}
\usage{
int32()

float64()

bool()

raw_col()

string_col()
}
\value{
A \code{shard_coltype} object.
}
\description{
Type constructors for schema-driven table outputs.
}
