% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shards.R
\name{create_contiguous_shards}
\alias{create_contiguous_shards}
\title{Create Contiguous Shards}
\usage{
create_contiguous_shards(n, block_size)
}
\arguments{
\item{n}{Total items.}

\item{block_size}{Items per shard.}
}
\value{
List of shard descriptors.
}
\description{
Creates shards with consecutive indices.
}
\keyword{internal}
