% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cow.R
\name{dim<-.shard_shared_vector}
\alias{dim<-.shard_shared_vector}
\title{Set dim on a Shared Vector}
\usage{
\method{dim}{shard_shared_vector}(x) <- value
}
\arguments{
\item{x}{A \code{shard_shared_vector}.}

\item{value}{Integer vector of dimensions.}
}
\value{
The modified object \code{x}.
}
\description{
Raises an error if the copy-on-write policy is \code{"deny"}.
}
