% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer.R
\name{dim.shard_buffer}
\alias{dim.shard_buffer}
\title{Dimensions of a Shared Memory Buffer}
\usage{
\method{dim}{shard_buffer}(x)
}
\arguments{
\item{x}{A \code{shard_buffer} object.}
}
\value{
An integer vector of dimensions, or \code{NULL} for 1-D buffers.
}
\description{
Dimensions of a Shared Memory Buffer
}
\examples{
\donttest{
buf <- buffer("double", dim = c(4, 5))
dim(buf)
buffer_close(buf)
}
}
