% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cow.R
\name{dimnames<-.shard_shared_vector}
\alias{dimnames<-.shard_shared_vector}
\title{Set dimnames on a Shared Vector}
\usage{
\method{dimnames}{shard_shared_vector}(x) <- value
}
\arguments{
\item{x}{A \code{shard_shared_vector}.}

\item{value}{List of dimnames.}
}
\value{
The modified object \code{x}.
}
\description{
Raises an error if the copy-on-write policy is \code{"deny"}.
}
