% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arena.R
\name{.arena_pop}
\alias{.arena_pop}
\title{Pop an arena from the stack}
\usage{
.arena_pop(arena_id, run_gc = FALSE)
}
\arguments{
\item{arena_id}{The ID to verify we're popping the right arena}

\item{run_gc}{Whether to trigger GC after pop}
}
\description{
Pop an arena from the stack
}
\keyword{internal}
