% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arena.R
\name{.estimate_object_size}
\alias{.estimate_object_size}
\title{Estimate object size safely}
\usage{
.estimate_object_size(x)
}
\arguments{
\item{x}{Object to measure}
}
\value{
Size in bytes or NA if cannot determine
}
\description{
Estimate object size safely
}
\keyword{internal}
