% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shard_map.R
\name{export_out_to_workers}
\alias{export_out_to_workers}
\title{Export Output Buffers to Workers}
\usage{
export_out_to_workers(pool, out)
}
\arguments{
\item{pool}{Worker pool.}

\item{out}{List of output buffers.}
}
\description{
Exports output buffer references to workers.
}
\keyword{internal}
