% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{factor_col}
\alias{factor_col}
\title{Categorical column type}
\usage{
factor_col(levels)
}
\arguments{
\item{levels}{Character vector of allowed levels.}
}
\value{
A \code{shard_coltype} object.
}
\description{
Stores factors as int32 codes plus shared levels metadata.
}
