% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shard_map.R
\name{get_profile_settings}
\alias{get_profile_settings}
\title{Get Profile Settings}
\usage{
get_profile_settings(profile, mem_cap, recycle)
}
\arguments{
\item{profile}{Profile name.}

\item{mem_cap}{User-specified memory cap.}

\item{recycle}{User-specified recycle setting.}
}
\value{
List of settings.
}
\description{
Returns settings based on execution profile.
}
\keyword{internal}
