% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arena.R
\name{in_arena}
\alias{in_arena}
\title{Check if Currently Inside an Arena}
\usage{
in_arena()
}
\value{
Logical indicating whether we are in an arena scope.
}
\description{
Returns TRUE if the current execution context is within an arena() scope.
}
\examples{
in_arena()
arena({
  in_arena()
})
}
