% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/share.R
\name{is_shared}
\alias{is_shared}
\title{Check if Object is Shared}
\usage{
is_shared(x)
}
\arguments{
\item{x}{An object to check.}
}
\value{
A logical scalar: \code{TRUE} if \code{x} is a shared object,
\code{FALSE} otherwise.
}
\description{
Check if Object is Shared
}
\examples{
is_shared(1:10)
\donttest{
shared <- share(1:10)
is_shared(shared)
close(shared)
}
}
