% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altrep.R
\name{is_shared_vector}
\alias{is_shared_vector}
\title{Check if an object is a shared vector}
\usage{
is_shared_vector(x)
}
\arguments{
\item{x}{Any R object}
}
\value{
TRUE if x is a shard ALTREP vector, FALSE otherwise
}
\description{
Check if an object is a shared vector
}
\examples{
\donttest{
seg <- segment_create(400)
segment_write(seg, 1:100, offset = 0)
x <- shared_vector(seg, "integer", length = 100)

is_shared_vector(x)
is_shared_vector(1:10)
}
}
