% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{is_view}
\alias{is_view}
\alias{is_block_view}
\title{View Predicates}
\usage{
is_view(x)

is_block_view(x)
}
\arguments{
\item{x}{An object.}
}
\value{
Logical. \code{TRUE} if \code{x} is a shard view (or block view).
}
\description{
View Predicates
}
\examples{
\donttest{
m <- share(matrix(1:20, nrow = 4))
v <- view_block(m, cols = idx_range(1, 2))
is_view(v)
is_block_view(v)
}
}
