% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{is_windows}
\alias{is_windows}
\title{Check if running on Windows}
\usage{
is_windows()
}
\value{
A logical scalar: \code{TRUE} if running on Windows, \code{FALSE} otherwise.
}
\description{
Check if running on Windows
}
\examples{
is_windows()
}
