% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer.R
\name{length.shard_buffer}
\alias{length.shard_buffer}
\title{Length of a Shared Memory Buffer}
\usage{
\method{length}{shard_buffer}(x)
}
\arguments{
\item{x}{A \code{shard_buffer} object.}
}
\value{
An integer scalar giving the total number of elements.
}
\description{
Length of a Shared Memory Buffer
}
\examples{
\donttest{
buf <- buffer("double", dim = 20)
length(buf)
buffer_close(buf)
}
}
