% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shards.R
\name{length.shard_descriptor}
\alias{length.shard_descriptor}
\title{Length of a shard_descriptor Object}
\usage{
\method{length}{shard_descriptor}(x)
}
\arguments{
\item{x}{A \code{shard_descriptor} object.}
}
\value{
An integer scalar giving the number of shards.
}
\description{
Length of a shard_descriptor Object
}
\examples{
sh <- shards(100, block_size = 25)
length(sh)
}
