% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shards_lazy.R
\name{length.shard_descriptor_lazy}
\alias{length.shard_descriptor_lazy}
\title{Length of a shard_descriptor_lazy Object}
\usage{
\method{length}{shard_descriptor_lazy}(x)
}
\arguments{
\item{x}{A \code{shard_descriptor_lazy} object.}
}
\value{
An integer scalar giving the number of shards.
}
\description{
Length of a shard_descriptor_lazy Object
}
\examples{
sh <- shards(100, block_size = 25)
length(sh)
}
