% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_registry.R
\name{list_kernels}
\alias{list_kernels}
\title{List registered kernels}
\usage{
list_kernels()
}
\value{
A character vector of registered kernel names.
}
\description{
List registered kernels
}
\examples{
list_kernels()
}
