% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shard_map.R
\name{make_chunk_executor}
\alias{make_chunk_executor}
\title{Create Chunk Executor Function}
\usage{
make_chunk_executor(auto_table = FALSE)
}
\value{
A function that executes chunks.
}
\description{
Creates a self-contained function that can execute a chunk in a worker.
This function is passed to dispatch_chunks and runs entirely within
the worker process.
}
\keyword{internal}
