% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/share.R
\name{materialize}
\alias{materialize}
\alias{materialize.shard_shared}
\alias{materialize.default}
\title{Materialize Shared Object}
\usage{
materialize(x)

\method{materialize}{shard_shared}(x)

\method{materialize}{default}(x)
}
\arguments{
\item{x}{A \code{shard_shared} object.}
}
\value{
The original R object.
}
\description{
Alias for \code{fetch()}. Retrieves the R object from shared memory.
}
\examples{
\donttest{
shared <- share(1:100)
data <- materialize(shared)
close(shared)
}
}
