% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{materialize.shard_view_block}
\alias{materialize.shard_view_block}
\title{Materialize a block view into an R matrix}
\usage{
\method{materialize}{shard_view_block}(x)
}
\arguments{
\item{x}{A \code{shard_view_block} object.}
}
\value{
A standard R matrix containing the selected rows and columns.
}
\description{
Materialize a block view into an R matrix
}
