% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{materialize.shard_view_gather}
\alias{materialize.shard_view_gather}
\title{Materialize a gather view into an R matrix}
\usage{
\method{materialize}{shard_view_gather}(x)
}
\arguments{
\item{x}{A \code{shard_view_gather} object.}
}
\value{
A standard R matrix containing the gathered columns.
}
\description{
Materialize a gather view into an R matrix
}
