% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cow.R
\name{names<-.shard_shared_vector}
\alias{names<-.shard_shared_vector}
\title{Set Names on a Shared Vector}
\usage{
\method{names}{shard_shared_vector}(x) <- value
}
\arguments{
\item{x}{A \code{shard_shared_vector}.}

\item{value}{Character vector of names.}
}
\value{
The modified object \code{x}.
}
\description{
Raises an error if the copy-on-write policy is \code{"deny"}.
}
