% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shards.R
\name{parse_count}
\alias{parse_count}
\title{Parse Count String}
\usage{
parse_count(x)
}
\arguments{
\item{x}{Character string.}
}
\value{
Integer value.
}
\description{
Parses strings like "1K", "10K", "1M" to integers.
}
\keyword{internal}
