% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker.R
\name{pid_is_alive}
\alias{pid_is_alive}
\title{Check if PID is Alive}
\usage{
pid_is_alive(pid)
}
\arguments{
\item{pid}{Process ID to check.}
}
\value{
Logical.
}
\description{
Check if PID is Alive
}
\keyword{internal}
