% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pkg_available}
\alias{pkg_available}
\title{Check Package Availability}
\usage{
pkg_available(pkg)
}
\arguments{
\item{pkg}{Package name.}
}
\value{
Logical. TRUE if package is available.
}
\description{
Check Package Availability
}
\keyword{internal}
