% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool.R
\name{pool_get}
\alias{pool_get}
\title{Get the Current Worker Pool}
\usage{
pool_get()
}
\value{
A \code{shard_pool} object or NULL.
}
\description{
Returns the active worker pool, or NULL if none exists.
}
\examples{
p <- pool_get()
is.null(p)
}
