% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arena.R
\name{print.arena_result}
\alias{print.arena_result}
\title{Print an arena_result object}
\usage{
\method{print}{arena_result}(x, ...)
}
\arguments{
\item{x}{An \code{arena_result} object.}

\item{...}{Additional arguments passed to \code{print}.}
}
\value{
Returns \code{x} invisibly.
}
\description{
Print an arena_result object
}
\examples{
\donttest{
info <- arena({ sum(1:10) }, diagnostics = TRUE)
print(info)
}
}
