% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergonomics.R
\name{print.shard_apply_policy}
\alias{print.shard_apply_policy}
\title{Print a shard_apply_policy Object}
\usage{
\method{print}{shard_apply_policy}(x, ...)
}
\arguments{
\item{x}{A \code{shard_apply_policy} object.}

\item{...}{Ignored.}
}
\value{
The input \code{x}, invisibly.
}
\description{
Print a shard_apply_policy Object
}
