% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer.R
\name{print.shard_buffer}
\alias{print.shard_buffer}
\title{Print a Shared Memory Buffer}
\usage{
\method{print}{shard_buffer}(x, ...)
}
\arguments{
\item{x}{A \code{shard_buffer} object.}

\item{...}{Ignored.}
}
\value{
The input \code{x}, invisibly.
}
\description{
Print a Shared Memory Buffer
}
\examples{
\donttest{
buf <- buffer("double", dim = 10)
print(buf)
buffer_close(buf)
}
}
