% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/share.R
\name{print.shard_deep_shared}
\alias{print.shard_deep_shared}
\title{Print a Deep-Shared Object}
\usage{
\method{print}{shard_deep_shared}(x, ...)
}
\arguments{
\item{x}{A \code{shard_deep_shared} object.}

\item{...}{Ignored.}
}
\value{
The input \code{x}, invisibly.
}
\description{
Print a Deep-Shared Object
}
\examples{
\donttest{
lst <- list(a = 1:10, b = 11:20)
shared <- share(lst, deep = TRUE, min_bytes = 1)
print(shared)
close(shared)
}
}
