% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shards.R
\name{print.shard_descriptor}
\alias{print.shard_descriptor}
\title{Print a shard_descriptor Object}
\usage{
\method{print}{shard_descriptor}(x, ...)
}
\arguments{
\item{x}{A \code{shard_descriptor} object.}

\item{...}{Further arguments (ignored).}
}
\value{
The input \code{x}, invisibly.
}
\description{
Print a shard_descriptor Object
}
\examples{
sh <- shards(100, block_size = 25)
print(sh)
}
