% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shards_lazy.R
\name{print.shard_descriptor_lazy}
\alias{print.shard_descriptor_lazy}
\title{Print a shard_descriptor_lazy Object}
\usage{
\method{print}{shard_descriptor_lazy}(x, ...)
}
\arguments{
\item{x}{A \code{shard_descriptor_lazy} object.}

\item{...}{Further arguments (ignored).}
}
\value{
The input \code{x}, invisibly.
}
\description{
Print a shard_descriptor_lazy Object
}
\examples{
sh <- shards(100, block_size = 25)
print(sh)
}
