% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool.R
\name{print.shard_health_report}
\alias{print.shard_health_report}
\title{Print a shard_health_report Object}
\usage{
\method{print}{shard_health_report}(x, ...)
}
\arguments{
\item{x}{A \code{shard_health_report} object.}

\item{...}{Further arguments (ignored).}
}
\value{
The input \code{x}, invisibly.
}
\description{
Print a shard_health_report Object
}
\examples{
\donttest{
p <- pool_create(2)
r <- pool_health_check(p)
print(r)
pool_stop(p)
}
}
