% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{print.shard_idx_range}
\alias{print.shard_idx_range}
\title{Print a shard_idx_range object}
\usage{
\method{print}{shard_idx_range}(x, ...)
}
\arguments{
\item{x}{A \code{shard_idx_range} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Returns \code{x} invisibly.
}
\description{
Print a shard_idx_range object
}
\examples{
r <- idx_range(1, 10)
print(r)
}
