% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool.R
\name{print.shard_pool}
\alias{print.shard_pool}
\title{Print a shard_pool Object}
\usage{
\method{print}{shard_pool}(x, ...)
}
\arguments{
\item{x}{A \code{shard_pool} object.}

\item{...}{Further arguments (ignored).}
}
\value{
The input \code{x}, invisibly.
}
\description{
Print a shard_pool Object
}
\examples{
\donttest{
p <- pool_create(2)
print(p)
pool_stop(p)
}
}
