% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shard_reduce.R
\name{print.shard_reduce_result}
\alias{print.shard_reduce_result}
\title{Print a shard_reduce_result Object}
\usage{
\method{print}{shard_reduce_result}(x, ...)
}
\arguments{
\item{x}{A \code{shard_reduce_result} object.}

\item{...}{Further arguments (ignored).}
}
\value{
The input \code{x}, invisibly.
}
\description{
Print a shard_reduce_result Object
}
\examples{
\donttest{
res <- shard_reduce(4L, map = function(s) sum(s$idx),
  combine = `+`, init = 0, workers = 2)
pool_stop()
print(res)
}
}
