% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shard_map.R
\name{print.shard_result}
\alias{print.shard_result}
\title{Print a shard_result Object}
\usage{
\method{print}{shard_result}(x, ...)
}
\arguments{
\item{x}{A \code{shard_result} object.}

\item{...}{Further arguments (ignored).}
}
\value{
The input \code{x}, invisibly.
}
\description{
Print a shard_result Object
}
\examples{
\donttest{
result <- shard_map(4L, function(shard) shard$idx, workers = 2)
pool_stop()
print(result)
}
}
