% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{print.shard_segment}
\alias{print.shard_segment}
\title{Print a Shared Memory Segment}
\usage{
\method{print}{shard_segment}(x, ...)
}
\arguments{
\item{x}{A \code{shard_segment} object.}

\item{...}{Ignored.}
}
\value{
The input \code{x}, invisibly.
}
\description{
Print a Shared Memory Segment
}
\examples{
\donttest{
seg <- segment_create(1024)
print(seg)
segment_close(seg)
}
}
