% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/share.R
\name{print.shard_shared}
\alias{print.shard_shared}
\title{Print a Shared Object}
\usage{
\method{print}{shard_shared}(x, ...)
}
\arguments{
\item{x}{A \code{shard_shared} object.}

\item{...}{Ignored.}
}
\value{
The input \code{x}, invisibly.
}
\description{
Print a Shared Object
}
\examples{
\donttest{
shared <- share(1:10)
print(shared)
close(shared)
}
}
