% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cow.R
\name{print.shard_shared_vector}
\alias{print.shard_shared_vector}
\title{Print a Shared Vector}
\usage{
\method{print}{shard_shared_vector}(x, ...)
}
\arguments{
\item{x}{A \code{shard_shared_vector}.}

\item{...}{Additional arguments passed to \code{print}.}
}
\value{
The input \code{x}, invisibly.
}
\description{
Print method for \code{shard_shared_vector} objects. Drops the wrapper class
and delegates to the underlying R print method.
}
