% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{print.shard_tiles}
\alias{print.shard_tiles}
\title{Print a shard_tiles object}
\usage{
\method{print}{shard_tiles}(x, ...)
}
\arguments{
\item{x}{A \code{shard_tiles} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Returns \code{x} invisibly.
}
\description{
Print a shard_tiles object
}
