% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{print.shard_view_block}
\alias{print.shard_view_block}
\title{Print a shard_view_block object}
\usage{
\method{print}{shard_view_block}(x, ...)
}
\arguments{
\item{x}{A \code{shard_view_block} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Returns \code{x} invisibly.
}
\description{
Print a shard_view_block object
}
