% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker.R
\name{print.shard_worker}
\alias{print.shard_worker}
\title{Print a shard_worker Object}
\usage{
\method{print}{shard_worker}(x, ...)
}
\arguments{
\item{x}{A \code{shard_worker} object.}

\item{...}{Further arguments (ignored).}
}
\value{
The input \code{x}, invisibly.
}
\description{
Print a shard_worker Object
}
\examples{
\donttest{
p <- pool_create(1)
print(p$workers[[1]])
pool_stop(p)
}
}
