% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{print_copy_report}
\alias{print_copy_report}
\title{Print Copy Report}
\usage{
print_copy_report(x)
}
\arguments{
\item{x}{A shard_report object with type "copy".}
}
\description{
Print Copy Report
}
\keyword{internal}
