% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{print_cow_report}
\alias{print_cow_report}
\title{Print COW Report}
\usage{
print_cow_report(x)
}
\arguments{
\item{x}{A shard_report object with type "cow".}
}
\description{
Print COW Report
}
\keyword{internal}
