% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{print_main_report}
\alias{print_main_report}
\title{Print Main Report}
\usage{
print_main_report(x)
}
\arguments{
\item{x}{A shard_report object.}
}
\description{
Print Main Report
}
\keyword{internal}
