% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{print_mem_report}
\alias{print_mem_report}
\title{Print Memory Report}
\usage{
print_mem_report(x)
}
\arguments{
\item{x}{A shard_report object with type "memory".}
}
\description{
Print Memory Report
}
\keyword{internal}
