% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{print_segment_report}
\alias{print_segment_report}
\title{Print Segment Report}
\usage{
print_segment_report(x)
}
\arguments{
\item{x}{A shard_report object with type "segment".}
}
\description{
Print Segment Report
}
\keyword{internal}
