% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{print_task_report}
\alias{print_task_report}
\title{Print Task Report}
\usage{
print_task_report(x)
}
\arguments{
\item{x}{A shard_report object with type "task".}
}
\description{
Print Task Report
}
\keyword{internal}
