% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue.R
\name{queue_complete}
\alias{queue_complete}
\title{Mark Chunk as Completed}
\usage{
queue_complete(queue, chunk_id, result = NULL)
}
\arguments{
\item{queue}{A \code{shard_queue} object.}

\item{chunk_id}{Chunk identifier.}

\item{result}{The result from processing.}
}
\value{
NULL (invisibly).
}
\description{
Mark Chunk as Completed
}
\keyword{internal}
