% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue.R
\name{queue_fail}
\alias{queue_fail}
\title{Mark Chunk as Failed}
\usage{
queue_fail(queue, chunk_id, error = NULL, requeue = TRUE)
}
\arguments{
\item{queue}{A \code{shard_queue} object.}

\item{chunk_id}{Chunk identifier.}

\item{error}{The error that occurred.}

\item{requeue}{Logical. Whether to requeue for retry.}
}
\value{
NULL (invisibly).
}
\description{
Mark Chunk as Failed
}
\keyword{internal}
