% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue.R
\name{queue_failures}
\alias{queue_failures}
\title{Get Failed Chunks}
\usage{
queue_failures(queue)
}
\arguments{
\item{queue}{A \code{shard_queue} object.}
}
\value{
List of failed chunk descriptors.
}
\description{
Get Failed Chunks
}
\keyword{internal}
