% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue.R
\name{queue_has_pending}
\alias{queue_has_pending}
\title{Check if Queue Has Work}
\usage{
queue_has_pending(queue)
}
\arguments{
\item{queue}{A \code{shard_queue} object.}
}
\value{
Logical. TRUE if there are pending chunks.
}
\description{
Check if Queue Has Work
}
\keyword{internal}
