% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue.R
\name{queue_is_done}
\alias{queue_is_done}
\title{Check if Queue is Done}
\usage{
queue_is_done(queue)
}
\arguments{
\item{queue}{A \code{shard_queue} object.}
}
\value{
Logical. TRUE if all chunks are completed or failed.
}
\description{
Check if Queue is Done
}
\keyword{internal}
