% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue.R
\name{queue_next}
\alias{queue_next}
\title{Get Next Chunk from Queue}
\usage{
queue_next(queue, worker_id)
}
\arguments{
\item{queue}{A \code{shard_queue} object.}

\item{worker_id}{Worker requesting the chunk.}
}
\value{
A chunk descriptor or NULL if queue is empty.
}
\description{
Get Next Chunk from Queue
}
\keyword{internal}
