% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue.R
\name{queue_requeue_worker}
\alias{queue_requeue_worker}
\title{Requeue All Chunks for a Worker}
\usage{
queue_requeue_worker(queue, worker_id)
}
\arguments{
\item{queue}{A \code{shard_queue} object.}

\item{worker_id}{Worker identifier.}
}
\value{
Integer. Number of chunks requeued.
}
\description{
When a worker dies or is recycled, requeue all its in-flight chunks.
}
\keyword{internal}
