% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue.R
\name{queue_results}
\alias{queue_results}
\title{Get Queue Results}
\usage{
queue_results(queue)
}
\arguments{
\item{queue}{A \code{shard_queue} object.}
}
\value{
List of results from completed chunks.
}
\description{
Get Queue Results
}
\keyword{internal}
