% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue.R
\name{queue_status}
\alias{queue_status}
\title{Get Queue Status}
\usage{
queue_status(queue)
}
\arguments{
\item{queue}{A \code{shard_queue} object.}
}
\value{
A list with queue statistics.
}
\description{
Get Queue Status
}
\keyword{internal}
