% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shard_map.R
\name{results}
\alias{results}
\title{Extract Results from shard_map}
\usage{
results(x, flatten = TRUE)
}
\arguments{
\item{x}{A shard_result object.}

\item{flatten}{Logical. Flatten nested results?}
}
\value{
List or vector of results.
}
\description{
Extract Results from shard_map
}
\examples{
\donttest{
result <- shard_map(4L, function(shard) shard$idx[[1L]], workers = 2)
pool_stop()
results(result)
}
}
