% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rss.R
\name{rss_via_proc}
\alias{rss_via_proc}
\title{Get RSS via /proc filesystem (Linux)}
\usage{
rss_via_proc(pid)
}
\arguments{
\item{pid}{Process ID.}
}
\value{
RSS in bytes or NA.
}
\description{
Get RSS via /proc filesystem (Linux)
}
\keyword{internal}
