% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rss.R
\name{rss_via_ps}
\alias{rss_via_ps}
\title{Get RSS via ps Package}
\usage{
rss_via_ps(pid)
}
\arguments{
\item{pid}{Process ID.}
}
\value{
RSS in bytes or NA.
}
\description{
Get RSS via ps Package
}
\keyword{internal}
