% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rss.R
\name{rss_via_ps_cmd}
\alias{rss_via_ps_cmd}
\title{Get RSS via ps Command (Unix)}
\usage{
rss_via_ps_cmd(pid)
}
\arguments{
\item{pid}{Process ID.}
}
\value{
RSS in bytes or NA.
}
\description{
Get RSS via ps Command (Unix)
}
\keyword{internal}
