% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rss.R
\name{rss_via_wmic}
\alias{rss_via_wmic}
\title{Get RSS via WMIC (Windows)}
\usage{
rss_via_wmic(pid)
}
\arguments{
\item{pid}{Process ID.}
}
\value{
RSS in bytes or NA.
}
\description{
Get RSS via WMIC (Windows)
}
\keyword{internal}
