% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{safe_div}
\alias{safe_div}
\title{Safe Divide}
\usage{
safe_div(x, y)
}
\arguments{
\item{x}{Numerator.}

\item{y}{Denominator.}
}
\value{
x/y or NA if y is zero.
}
\description{
Division that returns NA instead of Inf/NaN for zero divisors.
}
\keyword{internal}
