% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scratch.R
\name{scratch_diagnostics}
\alias{scratch_diagnostics}
\title{Scratch pool diagnostics}
\usage{
scratch_diagnostics()
}
\value{
A list with counters and current pool bytes.
}
\description{
Scratch pool diagnostics
}
\examples{
scratch_diagnostics()
}
