% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affinity.R
\name{segment_advise}
\alias{segment_advise}
\title{Advise OS about expected access pattern for a segment}
\usage{
segment_advise(
  seg,
  advice = c("normal", "sequential", "random", "willneed", "dontneed")
)
}
\arguments{
\item{seg}{A shard_segment.}

\item{advice}{One of "normal", "sequential", "random", "willneed", "dontneed".}
}
\value{
A logical scalar; \code{TRUE} if the OS accepted the hint.
}
\description{
This calls \code{madvise()} on the segment mapping when available.
}
\examples{
\donttest{
seg <- segment_create(1024)
segment_advise(seg, "sequential")
}
}
