% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{segment_close}
\alias{segment_close}
\title{Close a shared memory segment}
\usage{
segment_close(x, unlink = NULL)
}
\arguments{
\item{x}{A shard_segment object}

\item{unlink}{Whether to unlink the underlying file/shm (default: FALSE for
opened segments, TRUE for owned segments)}
}
\value{
\code{NULL}, invisibly.
}
\description{
Close a shared memory segment
}
\examples{
\donttest{
seg <- segment_create(1024)
segment_close(seg)
}
}
