% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{segment_path}
\alias{segment_path}
\title{Get the path or name of a segment}
\usage{
segment_path(x)
}
\arguments{
\item{x}{A shard_segment object}
}
\value{
The path string, or \code{NULL} for anonymous segments.
}
\description{
Get the path or name of a segment
}
\examples{
\donttest{
seg <- segment_create(1024, backing = "mmap")
segment_path(seg)
segment_close(seg)
}
}
