% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{segment_size}
\alias{segment_size}
\title{Get the size of a segment}
\usage{
segment_size(x)
}
\arguments{
\item{x}{A shard_segment object}
}
\value{
Size in bytes as a numeric scalar.
}
\description{
Get the size of a segment
}
\examples{
\donttest{
seg <- segment_create(1024)
segment_size(seg)
segment_close(seg)
}
}
