% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shard_map.R
\name{set_worker_seeds}
\alias{set_worker_seeds}
\title{Set Worker Seeds}
\usage{
set_worker_seeds(pool, seed, num_shards)
}
\arguments{
\item{pool}{Worker pool.}

\item{seed}{Base seed.}

\item{num_shards}{Number of shards for substream calculation.}
}
\description{
Sets reproducible RNG seeds in workers.
}
\keyword{internal}
