% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapter.R
\name{shard_get_adapter}
\alias{shard_get_adapter}
\title{Get Adapter for an Object}
\usage{
shard_get_adapter(x)
}
\arguments{
\item{x}{An R object.}
}
\value{
The adapter list if one is registered for any of the object's classes,
or NULL if no adapter is registered.
}
\description{
Retrieves the registered adapter for an object's class. Checks all classes
in the object's class hierarchy, returning the first matching adapter.
}
\examples{
shard_get_adapter(1:10)
}
