% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapter.R
\name{shard_list_adapters}
\alias{shard_list_adapters}
\title{List Registered Adapters}
\usage{
shard_list_adapters()
}
\value{
Character vector of class names with registered adapters.
}
\description{
Returns a character vector of all classes with registered adapters.
}
\examples{
shard_list_adapters()
}
