% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affinity.R
\name{shared_advise}
\alias{shared_advise}
\title{Advise access pattern for a shared input vector/matrix}
\usage{
shared_advise(
  x,
  advice = c("normal", "sequential", "random", "willneed", "dontneed")
)
}
\arguments{
\item{x}{A shard shared vector (from \code{\link[=share]{share()}}).}

\item{advice}{See \code{\link[=segment_advise]{segment_advise()}}.}
}
\value{
A logical scalar; \code{TRUE} if the OS accepted the hint.
}
\description{
Advise access pattern for a shared input vector/matrix
}
\examples{
\donttest{
x <- as_shared(1:100)
shared_advise(x, "sequential")
}
}
