% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/share.R
\name{shared_info}
\alias{shared_info}
\title{Get Information About a Shared Object}
\usage{
shared_info(x)
}
\arguments{
\item{x}{A \code{shard_shared} object.}
}
\value{
A named list with fields \code{path}, \code{backing}, \code{size},
\code{readonly}, \code{class_info}, and \code{segment_info}.
}
\description{
Get Information About a Shared Object
}
\examples{
\donttest{
shared <- share(1:100)
shared_info(shared)
close(shared)
}
}
