% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altrep.R
\name{shared_reset_diagnostics}
\alias{shared_reset_diagnostics}
\title{Reset diagnostic counters for a shared vector}
\usage{
shared_reset_diagnostics(x)
}
\arguments{
\item{x}{A shard ALTREP vector}
}
\value{
x (invisibly)
}
\description{
Reset diagnostic counters for a shared vector
}
\examples{
\donttest{
seg <- segment_create(400)
segment_write(seg, 1:100, offset = 0)
x <- shared_vector(seg, "integer", length = 100)

sum(x)
shared_diagnostics(x)$dataptr_calls

shared_reset_diagnostics(x)
shared_diagnostics(x)$dataptr_calls
}
}
