% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altrep.R
\name{shared_segment}
\alias{shared_segment}
\title{Get the underlying segment from a shared vector}
\usage{
shared_segment(x)
}
\arguments{
\item{x}{A shard ALTREP vector}
}
\value{
A \code{shard_segment} S3 object wrapping the underlying segment
}
\description{
Get the underlying segment from a shared vector
}
\examples{
\donttest{
x <- as_shared(1:100)
shared_segment(x)
}
}
