% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shards.R
\name{[.shard_descriptor}
\alias{[.shard_descriptor}
\title{Subset Shard Descriptor}
\usage{
\method{[}{shard_descriptor}(x, i)
}
\arguments{
\item{x}{A shard_descriptor object.}

\item{i}{Index or indices.}
}
\value{
A subset of the object.
}
\description{
Subset Shard Descriptor
}
\examples{
sh <- shards(100, block_size = 25)
sh[1:2]
}
