% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shards_lazy.R
\name{[.shard_descriptor_lazy}
\alias{[.shard_descriptor_lazy}
\title{Subset a shard_descriptor_lazy Object}
\usage{
\method{[}{shard_descriptor_lazy}(x, i)
}
\arguments{
\item{x}{A \code{shard_descriptor_lazy} object.}

\item{i}{Index or indices.}
}
\value{
A subset of the object.
}
\description{
Subset a shard_descriptor_lazy Object
}
\examples{
sh <- shards(100, block_size = 25)
sh[1:2]
}
