% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cow.R
\name{[[<-.shard_shared_vector}
\alias{[[<-.shard_shared_vector}
\title{Double-bracket Subset-assign a Shared Vector}
\usage{
\method{[[}{shard_shared_vector}(x, ...) <- value
}
\arguments{
\item{x}{A \code{shard_shared_vector}.}

\item{...}{Indices.}

\item{value}{Replacement value.}
}
\value{
The modified object \code{x}.
}
\description{
Replacement method for \code{shard_shared_vector}. Raises an error if the
copy-on-write policy is \code{"deny"}.
}
