% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer.R
\name{[<-.shard_buffer}
\alias{[<-.shard_buffer}
\title{Assign to Buffer Elements}
\usage{
\method{[}{shard_buffer}(x, i, j, ...) <- value
}
\arguments{
\item{x}{A shard_buffer object.}

\item{i}{Index or indices.}

\item{j}{Optional second index (for matrices).}

\item{...}{Additional indices (for arrays).}

\item{value}{Values to assign.}
}
\value{
The modified \code{shard_buffer} object, invisibly.
}
\description{
Assign to Buffer Elements
}
\examples{
\donttest{
buf <- buffer("double", dim = 10)
buf[1:5] <- rnorm(5)
buffer_close(buf)
}
}
