% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shard_map.R
\name{validate_borrow}
\alias{validate_borrow}
\title{Validate Borrowed Inputs}
\usage{
validate_borrow(borrow, cow)
}
\arguments{
\item{borrow}{List of inputs.}

\item{cow}{COW policy.}
}
\value{
Validated borrow list.
}
\description{
Validates and prepares borrowed inputs.
}
\keyword{internal}
