% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shard_map.R
\name{validate_out}
\alias{validate_out}
\title{Validate Output Buffers}
\usage{
validate_out(out)
}
\arguments{
\item{out}{List of output buffers.}
}
\value{
Validated out list.
}
\description{
Validates output buffer specifications.
}
\keyword{internal}
