% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{view}
\alias{view}
\title{Create a view over a shared matrix}
\usage{
view(x, rows = NULL, cols = NULL, type = c("auto", "block", "gather"))
}
\arguments{
\item{x}{A shared (share()d) atomic matrix (double/integer/logical/raw).}

\item{rows}{Row selector. NULL (all rows) or idx_range().}

\item{cols}{Column selector. NULL (all cols) or idx_range().}

\item{type}{View type. \code{"block"} or \code{"gather"} (or \code{"auto"}).}
}
\value{
A \code{shard_view_block} or \code{shard_view_gather} object depending
on the selectors provided.
}
\description{
Create a view over a shared matrix
}
\examples{
\donttest{
m <- share(matrix(1:20, nrow = 4))
v <- view(m, cols = idx_range(1, 2))
}
}
