% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{view_block}
\alias{view_block}
\title{Create a contiguous block view}
\usage{
view_block(x, rows = NULL, cols = NULL)
}
\arguments{
\item{x}{A shared (share()d) atomic matrix.}

\item{rows}{NULL or idx_range().}

\item{cols}{NULL or idx_range().}
}
\value{
A \code{shard_view_block} object representing the contiguous block slice.
}
\description{
Create a contiguous block view
}
\examples{
\donttest{
m <- share(matrix(1:20, nrow = 4))
v <- view_block(m, cols = idx_range(1, 2))
}
}
