% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{view_diagnostics}
\alias{view_diagnostics}
\title{View diagnostics}
\usage{
view_diagnostics()
}
\value{
A list with counters.
}
\description{
Returns global counters for view creation/materialization. This is a simple
first step; in future this should be integrated into shard_map run-level
diagnostics.
}
