% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker.R
\name{worker_eval}
\alias{worker_eval}
\title{Evaluate Expression in Worker}
\usage{
worker_eval(worker, expr, envir = parent.frame(), timeout = 3600)
}
\arguments{
\item{worker}{A \code{shard_worker} object.}

\item{expr}{Expression to evaluate.}

\item{envir}{Environment containing variables needed by expr.}

\item{timeout}{Numeric. Seconds to wait for result.}
}
\value{
The result of evaluation.
}
\description{
Sends an expression to the worker for evaluation.
}
\keyword{internal}
