% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker.R
\name{worker_is_alive}
\alias{worker_is_alive}
\title{Check if Worker is Alive}
\usage{
worker_is_alive(worker)
}
\arguments{
\item{worker}{A \code{shard_worker} object.}
}
\value{
Logical. TRUE if worker is alive.
}
\description{
Tests whether the worker process is still running.
}
\keyword{internal}
