% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker.R
\name{worker_metrics}
\alias{worker_metrics}
\title{Get Worker Metrics}
\usage{
worker_metrics(worker)
}
\arguments{
\item{worker}{A \code{shard_worker} object.}
}
\value{
A list of worker metrics.
}
\description{
Returns current metrics for a worker.
}
\keyword{internal}
