% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker.R
\name{worker_recycle}
\alias{worker_recycle}
\title{Recycle a Worker}
\usage{
worker_recycle(worker, init_expr = NULL, packages = NULL)
}
\arguments{
\item{worker}{A \code{shard_worker} object.}

\item{init_expr}{Expression to evaluate on startup.}

\item{packages}{Character vector. Packages to load.}
}
\value{
A new \code{shard_worker} object.
}
\description{
Kills the current worker and spawns a fresh replacement.
The new worker inherits the same ID but has a fresh R process.
}
\keyword{internal}
