% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker.R
\name{worker_spawn}
\alias{worker_spawn}
\title{Spawn a Worker Process}
\usage{
worker_spawn(id, init_expr = NULL, packages = NULL)
}
\arguments{
\item{id}{Integer. Worker identifier.}

\item{init_expr}{Expression to evaluate on startup.}

\item{packages}{Character vector. Packages to load.}
}
\value{
A \code{shard_worker} object.
}
\description{
Creates a new R worker process using parallel sockets.
}
\keyword{internal}
