% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{as_tibble}
\alias{as_tibble}
\title{Materialize a shard table handle as a data.frame/tibble}
\usage{
as_tibble(x, max_bytes = 256 * 1024^2, ...)
}
\arguments{
\item{x}{A shard table object.}

\item{max_bytes}{Warn if estimated payload exceeds this threshold.}

\item{...}{Reserved for future extensions.}
}
\description{
Materialize a shard table handle as a data.frame/tibble
}
