% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer.R
\name{buffer_info}
\alias{buffer_info}
\title{Get Buffer Info}
\usage{
buffer_info(x)
}
\arguments{
\item{x}{A shard_buffer object.}
}
\value{
A list with buffer properties.
}
\description{
Returns information about a buffer.
}
