% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{collect}
\alias{collect}
\title{Collect a shard table into memory}
\usage{
collect(x, ...)
}
\arguments{
\item{x}{A shard table handle.}

\item{...}{Passed to \code{as_tibble()}.}
}
\description{
\code{collect()} is a convenience alias for \code{as_tibble()} for shard table outputs.
}
