% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/share.R
\name{is_shared}
\alias{is_shared}
\title{Check if Object is Shared}
\usage{
is_shared(x)
}
\arguments{
\item{x}{An object to check.}
}
\value{
TRUE if x is a \code{shard_shared} or \code{shard_deep_shared} object,
FALSE otherwise.
}
\description{
Check if Object is Shared
}
\examples{
is_shared(share(1:10))  # TRUE
is_shared(1:10)         # FALSE
}
