% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{is_view}
\alias{is_view}
\alias{is_block_view}
\title{View Predicates}
\usage{
is_view(x)

is_block_view(x)
}
\arguments{
\item{x}{An object.}
}
\value{
Logical. TRUE if \code{x} is a shard view (or block view).
}
\description{
View Predicates
}
