% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{segment_info}
\alias{segment_info}
\title{Get segment information}
\usage{
segment_info(x)
}
\arguments{
\item{x}{A shard_segment object}
}
\value{
A list with segment properties
}
\description{
Get segment information
}
