% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{segment_open}
\alias{segment_open}
\title{Open an existing shared memory segment}
\usage{
segment_open(path, backing = c("mmap", "shm"), readonly = TRUE)
}
\arguments{
\item{path}{Path or shm name of the segment}

\item{backing}{Backing type: "mmap" or "shm"}

\item{readonly}{Open as read-only}
}
\value{
An S3 object of class "shard_segment"
}
\description{
Open an existing shared memory segment
}
