% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{segment_protect}
\alias{segment_protect}
\title{Make a segment read-only}
\usage{
segment_protect(x)
}
\arguments{
\item{x}{A shard_segment object}
}
\value{
The segment object (invisibly)
}
\description{
Make a segment read-only
}
