% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{segment_read}
\alias{segment_read}
\title{Read raw data from a segment}
\usage{
segment_read(x, offset = 0, size = NULL)
}
\arguments{
\item{x}{A shard_segment object}

\item{offset}{Byte offset to start reading (0-based)}

\item{size}{Number of bytes to read}
}
\value{
A raw vector
}
\description{
Read raw data from a segment
}
