% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/share.R
\name{shared_info}
\alias{shared_info}
\title{Get Information About a Shared Object}
\usage{
shared_info(x)
}
\arguments{
\item{x}{A \code{shard_shared} object.}
}
\value{
A list with detailed information about the shared segment.
}
\description{
Get Information About a Shared Object
}
