% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_stream.R
\name{stream_count}
\alias{stream_count}
\title{Stream row count}
\usage{
stream_count(x)
}
\arguments{
\item{x}{A \code{shard_row_groups} or \code{shard_dataset} handle.}
}
\value{
Total number of rows.
}
\description{
Stream row count
}
