% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer.R
\name{[.shard_buffer}
\alias{[.shard_buffer}
\title{Extract Buffer Elements}
\usage{
\method{[}{shard_buffer}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{A shard_buffer object.}

\item{i}{Index or indices.}

\item{j}{Optional second index (for matrices).}

\item{...}{Additional indices (for arrays).}

\item{drop}{Whether to drop dimensions.}
}
\value{
Vector or array of values.
}
\description{
Extract Buffer Elements
}
