% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shards.R
\name{[[.shard_descriptor}
\alias{[[.shard_descriptor}
\title{Get Single Shard}
\usage{
\method{[[}{shard_descriptor}(x, i)
}
\arguments{
\item{x}{A shard_descriptor object.}

\item{i}{Index.}
}
\value{
The selected shard.
}
\description{
Get Single Shard
}
