% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shard_map.R
\name{succeeded}
\alias{succeeded}
\title{Check if shard_map Succeeded}
\usage{
succeeded(x)
}
\arguments{
\item{x}{A shard_result object.}
}
\value{
Logical. TRUE if no failures.
}
\description{
Check if shard_map Succeeded
}
