% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{view_info}
\alias{view_info}
\title{Introspection for a view}
\usage{
view_info(v)
}
\arguments{
\item{v}{A shard view.}
}
\value{
A named list of metadata.
}
\description{
Returns metadata about a view without forcing materialization.
}
