% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker.R
\name{worker_kill}
\alias{worker_kill}
\title{Kill a Worker Process}
\usage{
worker_kill(worker)
}
\arguments{
\item{worker}{A \code{shard_worker} object.}
}
\value{
NULL (invisibly).
}
\description{
Terminates the worker process and closes connections.
}
\keyword{internal}
